+++ package.json
@@ -24,149 +24,5 @@
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-plugin-html": "^6.2.0",
     "eslint-plugin-import": "^2.26.0"
-  },
-  "dependencies": {
-    "ajv": "^6.12.6",
-    "acorn-jsx": "^5.3.2",
-    "acorn": "^8.7.1",
-    "array-includes": "^3.1.5",
-    "array.prototype.flat": "^1.3.0",
-    "balanced-match": "^1.0.2",
-    "call-bind": "^1.0.2",
-    "brace-expansion": "^1.1.11",
-    "callsites": "^3.1.0",
-    "color-name": "^1.1.4",
-    "color-convert": "^2.0.1",
-    "confusing-browser-globals": "^1.0.11",
-    "cross-spawn": "^7.0.3",
-    "concat-map": "^0.0.1",
-    "doctrine": "^3.0.0",
-    "deep-is": "^0.1.4",
-    "chalk": "^4.1.2",
-    "debug": "^4.3.4",
-    "domhandler": "^4.3.1",
-    "domelementtype": "^2.3.0",
-    "dom-serializer": "^1.4.1",
-    "domutils": "^2.8.0",
-    "es-shim-unscopables": "^1.0.0",
-    "define-properties": "^1.1.4",
-    "ansi-styles": "^4.3.0",
-    "es-to-primitive": "^1.2.1",
-    "es-abstract": "^1.20.1",
-    "eslint-scope": "^7.1.1",
-    "escape-string-regexp": "^4.0.0",
-    "eslint-import-resolver-node": "^0.3.6",
-    "eslint-visitor-keys": "^3.3.0",
-    "entities": "^3.0.1",
-    "eslint-utils": "^3.0.0",
-    "espree": "^9.3.2",
-    "esquery": "^1.4.0",
-    "ansi-regex": "^5.0.1",
-    "esrecurse": "^4.3.0",
-    "eslint-module-utils": "^2.7.3",
-    "fast-deep-equal": "^3.1.3",
-    "estraverse": "^5.3.0",
-    "esutils": "^2.0.3",
-    "fast-levenshtein": "^2.0.6",
-    "fast-json-stable-stringify": "^2.1.0",
-    "find-up": "^2.1.0",
-    "function-bind": "^1.1.1",
-    "fs.realpath": "^1.0.0",
-    "function.prototype.name": "^1.1.5",
-    "file-entry-cache": "^6.0.1",
-    "functional-red-black-tree": "^1.0.1",
-    "flatted": "^3.2.6",
-    "argparse": "^2.0.1",
-    "eslint-config-prettier": "^8.5.0",
-    "get-intrinsic": "^1.1.2",
-    "get-symbol-description": "^1.0.0",
-    "has": "^1.0.3",
-    "glob": "^7.2.3",
-    "has-property-descriptors": "^1.0.0",
-    "glob-parent": "^6.0.2",
-    "has-flag": "^4.0.0",
-    "has-symbols": "^1.0.3",
-    "ignore": "^5.2.0",
-    "globals": "^13.15.0",
-    "htmlparser2": "^7.2.0",
-    "has-tostringtag": "^1.0.0",
-    "imurmurhash": "^0.1.4",
-    "inflight": "^1.0.6",
-    "import-fresh": "^3.3.0",
-    "has-bigints": "^1.0.2",
-    "internal-slot": "^1.0.3",
-    "inherits": "^2.0.4",
-    "flat-cache": "^3.0.4",
-    "is-extglob": "^2.1.1",
-    "is-date-object": "^1.0.5",
-    "is-boolean-object": "^1.1.2",
-    "is-callable": "^1.2.4",
-    "is-negative-zero": "^2.0.2",
-    "functions-have-names": "^1.2.3",
-    "is-core-module": "^2.9.0",
-    "is-string": "^1.0.7",
-    "is-regex": "^1.1.4",
-    "is-glob": "^4.0.3",
-    "is-bigint": "^1.0.4",
-    "isexe": "^2.0.0",
-    "json-schema-traverse": "^0.4.1",
-    "json-stable-stringify-without-jsonify": "^1.0.1",
-    "locate-path": "^2.0.0",
-    "lodash.merge": "^4.6.2",
-    "levn": "^0.4.1",
-    "is-symbol": "^1.0.4",
-    "json5": "^1.0.1",
-    "js-yaml": "^4.1.0",
-    "ms": "^2.1.2",
-    "natural-compare": "^1.4.0",
-    "minimist": "^1.2.6",
-    "is-number-object": "^1.0.7",
-    "object-inspect": "^1.12.2",
-    "object-keys": "^1.1.1",
-    "once": "^1.4.0",
-    "minimatch": "^3.1.2",
-    "p-limit": "^1.3.0",
-    "p-try": "^1.0.0",
-    "p-locate": "^2.0.0",
-    "object.entries": "^1.1.5",
-    "object.values": "^1.1.5",
-    "parent-module": "^1.0.1",
-    "optionator": "^0.9.1",
-    "path-is-absolute": "^1.0.1",
-    "is-shared-array-buffer": "^1.0.2",
-    "is-weakref": "^1.0.2",
-    "path-parse": "^1.0.7",
-    "prelude-ls": "^1.2.1",
-    "regexp.prototype.flags": "^1.4.3",
-    "path-exists": "^3.0.0",
-    "regexpp": "^3.2.0",
-    "semver": "^6.3.0",
-    "resolve-from": "^4.0.0",
-    "shebang-regex": "^3.0.0",
-    "punycode": "^2.1.1",
-    "rimraf": "^3.0.2",
-    "resolve": "^1.22.1",
-    "path-key": "^3.1.1",
-    "shebang-command": "^2.0.0",
-    "side-channel": "^1.0.4",
-    "string.prototype.trimend": "^1.0.5",
-    "strip-ansi": "^6.0.1",
-    "object.assign": "^4.1.2",
-    "strip-bom": "^3.0.0",
-    "supports-color": "^7.2.0",
-    "string.prototype.trimstart": "^1.0.5",
-    "strip-json-comments": "^3.1.1",
-    "text-table": "^0.2.0",
-    "tsconfig-paths": "^3.14.1",
-    "uri-js": "^4.4.1",
-    "type-check": "^0.4.0",
-    "which": "^2.0.2",
-    "type-fest": "^0.20.2",
-    "wrappy": "^1.0.2",
-    "which-boxed-primitive": "^1.0.2",
-    "supports-preserve-symlinks-flag": "^1.0.0",
-    "word-wrap": "^1.2.3",
-    "unbox-primitive": "^1.0.2",
-    "v8-compile-cache": "^2.3.0"
   }
 }
