+++ src/js/constants.js
@@ -1,0 +1,12 @@
+const MAX_INPUT_LENGTH = 3;
+
+const INITIAL_STATE = {
+  accumulator: 0,
+  number: null,
+  operator: '',
+};
+
+export {
+  MAX_INPUT_LENGTH,
+  INITIAL_STATE,
+};
